Added back the explicit del asyncio and other minor stuff
