+++ uvloop/includes/stdlib.pxi
@@ -41,6 +41,7 @@ cdef aio_SSLProtocol = asyncio.sslproto.SSLProtocol
 cdef aio_isfuture = getattr(asyncio, 'isfuture', None)
 cdef aio_get_running_loop = getattr(asyncio, '_get_running_loop', None)
 cdef aio_set_running_loop = getattr(asyncio, '_set_running_loop', None)
+cdef aio_debug_wrapper = getattr(asyncio.coroutines, 'debug_wrapper', None)
 
 cdef col_deque = collections.deque
 cdef col_Iterable = collections.Iterable
@@ -145,8 +146,8 @@ cdef py_inf = float('inf')
 
 
 # Cython doesn't clean-up imported objects properly in Py3 mode,
-# so we delete refs to all modules manually (except sys, asyncio)
-del concurrent, collections, errno
+# so we delete refs to all modules manually (except sys)
+del asyncio, concurrent, collections, errno
 del functools, inspect, itertools, socket, os, threading
 del signal, subprocess, ssl
 del time, traceback, warnings, weakref
